Make this stuff compile, it breaks lines a little too often, but otherwise we can continue on.
[git-p4: depot-paths = "//depot/qt/main/": change = 127241]

+++ src/gui/text/qfont_mac.cpp
@@ -78,7 +78,7 @@ int QFontMetrics::charWidth(const QString &str, int pos) const
 {
     QTextEngine layout(str, d);
     layout.itemize(QTextEngine::WidthOnly);
-    return layout.width(pos, 1);
+    return layout.width(pos, 1).toInt();
 }
 
 QString QFont::rawName() const
